% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DropRedundantFilings.R
\name{DropRedundantFilings}
\alias{DropRedundantFilings}
\title{DropRedundantFilings}
\usage{
DropRedundantFilings(df)
}
\arguments{
\item{df}{A data.frame containing a T1 sheet}
}
\value{
A data.frame without redundant filings
}
\description{
Drops redundant filings from a data.frame containing a T1 sheet. Redundancy is defined as multiple filings on the same date.
}

