% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FundamentalsInfo.R
\name{FundamentalsInfo}
\alias{FundamentalsInfo}
\title{FundamentalsInfo}
\usage{
FundamentalsInfo(fundamental.name = NULL)
}
\arguments{
\item{fundamental.name}{(optional) A name of a fundamental (e.g. Net Income). If none is specified, entire fundamentals matrix is returned.}
}
\value{
A matrix of information about one or more fundamentals.
}
\description{
Returns basic information about the location of various fundamentals within the T1 sheets
}

