% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FundCode.R
\name{FundCode}
\alias{FundCode}
\title{FundCode}
\usage{
FundCode(fundamental.names)
}
\arguments{
\item{fundamental.names}{A character vector of names of fundamentals (e.g. 'Total Revenue')}
}
\value{
A character vector of codes.
}
\description{
Maps names of common fundamentals to their four digit T1 codes
}

