% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetFirms.R
\name{GetFirms}
\alias{GetFirms}
\title{GetFirms}
\usage{
GetFirms(sheets)
}
\arguments{
\item{sheets}{A list of T1 excel sheets}
}
\value{
A character vector of firms whose financial statements are contained in sheets.
}
\description{
Returns a character vector of firms whose financial statements are contained in a list of sheets.
}

