% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetSheet.R
\name{GetSheet}
\alias{GetSheet}
\title{GetSheet}
\usage{
GetSheet(fundamental.info, sheets)
}
\arguments{
\item{fundamental.info}{A character vector of fundamental info, output from FundamentalsInfo.}

\item{sheets}{List of sheets}
}
\description{
Grabs a particular T1 sheet from a list of sheets that were previously read into memory.
}

