% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetStockPriceAverages.R
\name{GetStockPriceAverages}
\alias{GetStockPriceAverages}
\title{GetStockPriceAverages}
\usage{
GetStockPriceAverages(symbol, start.date, forward.dates = c(5, 10))
}
\arguments{
\item{symbol}{A ticker symbol (e.g. 'AAPL')}

\item{start.date}{Start date for analysis.}

\item{forward.dates}{A character vector of relative forward dates (e.g. 5 days from start.date, 10 days from start.date)}
}
\value{
A list of average stock prices
}
\description{
Retrieves the average stock price for a given ticker symbol across a series of dates.
}

