% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetStockTimeseries.R
\name{GetStockTimeseries}
\alias{GetStockTimeseries}
\title{GetStockTimeseries}
\usage{
GetStockTimeseries(symbol, from.date, to.date)
}
\arguments{
\item{symbol}{A ticker symbol}

\item{from.date}{Beginning date}

\item{to.date}{End date}
}
\value{
A data.frame of closing prices
}
\description{
GetStockTimeseries
}

