% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetStructuralParameters.R
\name{GetStructuralParameters}
\alias{GetStructuralParameters}
\title{GetStructuralParameters}
\usage{
GetStructuralParameters(sheet.type)
}
\arguments{
\item{sheet.type}{Type of sheet (i.e. Balance Sheet)}
}
\value{
A list of structural parameters depending on the sheet type.
}
\description{
GetStructuralParameters
}

