% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MakeAttributes.R
\name{MakeAttributes}
\alias{MakeAttributes}
\title{MakeAttributes}
\usage{
MakeAttributes(sheet.type, df)
}
\arguments{
\item{sheet.type}{Type of T1 sheet (e.g. Balance Sheet)}

\item{df}{Data.frame containing a T1 sheet}
}
\value{
A list of attributes
}
\description{
Produces a set of attributes for a T1 sheet data.frame, depending on its sheet type
}

