% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RatioInfo.R
\name{RatioInfo}
\alias{RatioInfo}
\title{RatioInfo}
\usage{
RatioInfo(ratio.name = NULL)
}
\arguments{
\item{ratio.name}{(optional) A ratio name (e.g. Earnings per Share) If none is specified, entire list is returned}
}
\value{
A list containing basic information about each ratio.
}
\description{
Returns basic information about how to calculate fundamentals ratios.
}

