% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadT1Directory.R
\name{ReadT1Directory}
\alias{ReadT1Directory}
\title{ReadT1Directory}
\usage{
ReadT1Directory(directory)
}
\arguments{
\item{directory}{Path to directory}
}
\value{
A list containing all of the data.frames that were read into memory.
}
\description{
Reads a directory of Thomson ONE excel output files (.xlsx format) into memory.
}

