% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadT1File.R
\name{ReadT1File}
\alias{ReadT1File}
\title{ReadT1File}
\usage{
ReadT1File(file.name)
}
\arguments{
\item{file.name}{Name of T1 excel/HTML file}
}
\value{
A data.frame containing the contents of the .xlsx file.
}
\description{
Read a Thomson ONE excel/HTML output file into memory
}

