% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/StripFormatting.R
\name{StripFormatting}
\alias{StripFormatting}
\title{StripFormatting}
\usage{
StripFormatting(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A modified character vector
}
\description{
Performs the following operations on a character input: removes commas, coverts numbers in parentheses to negative numbers, and removes trailing whitespace.
}

