% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/YearsInCommon.R
\name{YearsInCommon}
\alias{YearsInCommon}
\title{YearsInCommon}
\usage{
YearsInCommon(firm.name, fundamentals.names, sheets, years = NULL)
}
\arguments{
\item{firm.name}{Name of firm}

\item{fundamentals.names}{A character vector of fundamentals}

\item{sheets}{A list of T1 sheets}

\item{years}{(optional) A character vector of years.}
}
\value{
A character vector of matching years.
}
\description{
Finds the years in common between a vector of fundamentals. If years is specified, then the function will only check those years.
}

