% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DatesInCommon.R
\name{DatesInCommon}
\alias{DatesInCommon}
\title{DatesInCommon}
\usage{
DatesInCommon(firm.name, fundamentals.codes, sheets, dates = NULL)
}
\arguments{
\item{firm.name}{Name of firm}

\item{sheets}{A list of T1 sheets}

\item{dates}{(optional) A character vector of reporting dates. These can either be years (e.g. '2014') or specific dates (e.g. '2014-05-12).}

\item{fundamentals.names}{A character vector of fundamentals}
}
\value{
A character vector of matching dates.
}
\description{
Finds the reporting periods in common between a vector of fundamentals. If years is specified, then the function will only check those years.
}

