% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FundamentalsInfo.R
\name{FundamentalsInfo}
\alias{FundamentalsInfo}
\title{FundamentalsInfo}
\usage{
FundamentalsInfo(firm.name, fundamental.code, sheets)
}
\arguments{
\item{firm.name}{Name of firm}

\item{fundamental.code}{Four digit code for fundamental (e.g. 'NINC' for 'Net Income')}

\item{A}{list of T1 excel sheets}
}
\description{
Determines the sheet and row containing a fundamental for a particular firm, as well as its full name.
}

