% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetFundamentals.R
\name{GetFundamentals}
\alias{GetFundamentals}
\title{GetFundamentals}
\usage{
GetFundamentals(firm.name, fundamentals.names, sheets, years = NULL)
}
\arguments{
\item{firm.name}{Name of firm}

\item{fundamentals.names}{A character vector of fundamentals}

\item{sheets}{A list of T1 sheets}

\item{years}{(optional) A character vector of years}
}
\value{
A matrix of fundamentals (one per column)
}
\description{
Extracts a set of fundamentals for a given firm
}

