% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetRatios.R
\name{GetRatios}
\alias{GetRatios}
\title{GetRatio}
\usage{
GetRatios(firm.name, sheets, ratio.names = NULL, years = NULL)
}
\arguments{
\item{firm.name}{Name of firm}

\item{sheets}{A list of T1 sheets}

\item{ratio.names}{(optional) A character vector of ratio names. If none is specified, all available ratios are returned.}

\item{years}{(optional) A character vector of years.}
}
\value{
A matrix of ratios (one per column)
}
\description{
Calculates a set of fundamental ratios for a given firm
}

