% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GetSheet.R
\name{GetSheet}
\alias{GetSheet}
\title{GetSheet}
\usage{
GetSheet(firm.name, sheet.type, sheets)
}
\arguments{
\item{firm.name}{Name of the firm}

\item{sheet.type}{Type of sheet (e.g. Balance Sheet)}

\item{sheets}{List of sheets}
}
\description{
Grabs a particular T1 sheet from a list of sheets that were previously read into memory.
}

