% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReadT1File.R
\name{ReadT1File}
\alias{ReadT1File}
\title{ReadT1File}
\usage{
ReadT1File(file.name, sheet.types = c("Balance Sheet", "Cash Flow Statement",
  "Income Statement"))
}
\arguments{
\item{file.name}{Name of .xlsx file}

\item{sheet.types}{A character vector of sheet types to look for.}
}
\value{
A data.frame containing the contents of the .xlsx file.
}
\description{
Read a Thomson ONE excel output file (.xlsx format) into memory
}

